Fix schema dumper behavior about datetime precision value
By #42297, the default value of datetime precision in rails is now 6.
With this change, if the value of datetime precision is 0, it should be
dumped into schema, but it is not.
So I modified it to dump to schema when the value of datetime precision
is 0, and not to dump when the value is 6, which is the default value of
rails.