Split AR Encryption config settings into sections [ci-skip]
This splits the table of `config.active_record.encryption.*` settings
into individual linkable sections, similar to the Configuration guide.

This also bumps up the advice about using credentials, adds a few
backticks in the appropriate spots, and fixes a link which always
pointed to the latest version of the Configuration guide.