Use tabGroups as a proxy for tags
Firefox has a useful feature of tagging bookmarks. Chrome (and Brave)
have better support for tab groups. It would be useful to develop a
proxy for these two features.

Using Downstream database as a source of truth, browser profiles can be
"printed" and shared with 3rd parties.

This could serve as a supervised learning tool for ML platforms, with
troves of data being added to the profile to enhance learning capacity
almost like a speed reading reference gathering tool.