Starting the process of unified tagging interface
You can use tabGroups to add bookmarks and tag them by updating the
query string to use the tab group name (and anything else that can be
derived from the webpage).

Like I've said before, chrome and Brave have tabGroups, firefox has tags
and keywords. There's gott be a way to make them talk nice.

The downstream API should be able to handle these cases. When receiving
payloads with tags in the URL as per query string syntax, or from the
data in the bookmark itself when coming from firefox.